Add apps-env subpart of the project. This part contains all the source code of the execution middleware applications. You do not need to compile these applications, they are already compiled for arm architecture and packaged in binRelease/sdImage_pkg.bin
If you still want to compile them, you should install an arm compiler
and update the toolchain file in cmake-misc.
Furthermore, for the driver compilation, we should get the linux
source from a petalinux project.
A make target is available in qemu-env to help the user creating a new
petalinux project.